pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openbox-deepagent-sdk-python"
version = "0.1.0"
description = "OpenBox governance SDK for DeepAgents (langchain-ai/deepagents)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "openbox-langgraph-sdk-python>=0.1.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
]

[project.optional-dependencies]
deepagents = [
    "deepagents>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["openbox_deepagent"]


[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "PIE", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
